pyproject.toml
[project]
name = "pylox"
version = "0.1.0"
description = "pylox interpreter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
]
